move retry to pkg (#631)
