+++ pkg/retry/retry.go
@@ -13,6 +13,10 @@
 // limitations under the License.
 
 // Package retry provides generic function-based retry helpers.
+//
+// Although exported, this package is non intended for general consumption.
+// It is a shared dependency between multiple exposure notifications projects.
+// We cannot guarantee that there won't be breaking changes in the future.
 package retry
 
 import (
